test(stringcalculator) : int -> BigInteger 변경에 따른 테스트 케이스 수정

+++ src/test/kotlin/calculator/model/StringCalculatorTest.kt
@@ -9,8 +9,6 @@ class StringCalculatorTest {
     fun addNumbersTest() {
         val calculator = StringCalculator()
         val num = calculator.addNumbers(listOf("1", "105"))
-        assertEquals(106, num)
+        assertEquals(106.toBigInteger(), num)
     }
-
-
 }
